staying current
fixed warnings in new Xcode version, etc.

+++ ScoreTracker/PlayerControls.m
@@ -14,59 +14,59 @@
 
 - (id)initWithNibName:(NSString *)nibNameOrNil bundle:(NSBundle *)nibBundleOrNil
 {
-    self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
-    if (self) {
-        // Custom initialization
-    }
-    return self;
+	self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
+	if (self) {
+		// Custom initialization
+	}
+	return self;
 }
 
 - (void)didReceiveMemoryWarning
 {
-    // Releases the view if it doesn't have a superview.
-    [super didReceiveMemoryWarning];
-    
-    // Release any cached data, images, etc that aren't in use.
+	// Releases the view if it doesn't have a superview.
+	[super didReceiveMemoryWarning];
+
+	// Release any cached data, images, etc that aren't in use.
 }
 
 #pragma mark - View lifecycle
 
 - (void)viewDidLoad
 {
-    [super viewDidLoad];
-    // Do any additional setup after loading the view from its nib.
+	[super viewDidLoad];
+	// Do any additional setup after loading the view from its nib.
 }
 
--(void)viewWillAppear:(BOOL)animated
+- (void)viewWillAppear:(BOOL)animated
 {
-    Score = 0;
+	Score = 0;
 }
 
 - (void)viewDidUnload
 {
-    [self setPlayname:nil];
-    [self setPlayScore:nil];
-    [super viewDidUnload];
-    // Release any retained subviews of the main view.
-    // e.g. self.myOutlet = nil;
+	[self setPlayname:nil];
+	[self setPlayScore:nil];
+	[super viewDidUnload];
+	// Release any retained subviews of the main view.
+	// e.g. self.myOutlet = nil;
 }
 
 - (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)interfaceOrientation
 {
-    // Return YES for supported orientations
-    return (interfaceOrientation == UIInterfaceOrientationPortrait);
+	// Return YES for supported orientations
+	return (interfaceOrientation == UIInterfaceOrientationPortrait);
 }
 
-- (IBAction)pressedScorePlus:(id)sender 
+- (IBAction)pressedScorePlus:(id)sender
 {
-    Score++;
-    playScore.text = [NSString stringWithFormat:@"%d", Score];
+	Score++;
+	playScore.text = [NSString stringWithFormat:@"%ld", (long)Score];
 }
 
-- (IBAction)pressedScoreMinus:(id)sender 
+- (IBAction)pressedScoreMinus:(id)sender
 {
-    Score--;
-    playScore.text = [NSString stringWithFormat:@"%d", Score];
+	Score--;
+	playScore.text = [NSString stringWithFormat:@"%ld", Score];
 }
 
 @end
